Continue listening for media buttons while paused
After being in the background and paused for a minute or so, Android 8
and up log "ActivityManager: Stopping service due to app idle" and cause
MusicService.onDestroy to be called. If we release the
MediaSessionCompat at this point, Shuttle will no longer respond to
media buttons, such as a headset play/pause button, so we need to only
release this if the app is no longer needed, maybe when it's been swiped
out of recents and the notification has been dismissed.

By not cleaning this up, will resources be leaked? Do we need to find
and release the previous instance if the service starts again?

See #391